fix text coverage and refactor attrs/cm utils
