+++ dascore/__init__.py
@@ -5,6 +5,8 @@ from rich import print  # noqa
 from dascore.core.patch import Patch
 from dascore.core.attrs import PatchAttrs
 from dascore.core.spool import BaseSpool, spool
+from dascore.core.coordmanager import get_coord_manager, CoordManager
+from dascore.core.coords import get_coord
 from dascore.examples import get_example_patch, get_example_spool
 from dascore.io.core import get_format, read, scan, scan_to_df, write
 from dascore.units import get_quantity, get_unit
