+++ dascore/core/attrs.py
@@ -2,40 +2,24 @@
 from __future__ import annotations
 
 import warnings
-from collections import ChainMap, defaultdict
 from collections.abc import Mapping, Sequence
-from functools import reduce
 from typing import Annotated, Any, Literal
 
 import numpy as np
-import pandas as pd
 from pydantic import ConfigDict, Field, PlainValidator, model_validator
 from typing_extensions import Self
 
 import dascore as dc
 from dascore.constants import (
-    DEFAULT_ATTRS_TO_IGNORE,
     VALID_DATA_CATEGORIES,
     VALID_DATA_TYPES,
-    WARN_LEVELS,
-    PatchType,
-    attr_conflict_description,
     max_lens,
 )
-from dascore.core.coordmanager import CoordManager
 from dascore.core.coords import BaseCoord, CoordSummary
-from dascore.exceptions import AttributeMergeError, IncompatiblePatchError
-from dascore.utils.docs import compose_docstring
+from dascore.utils.attrs import separate_coord_info
 from dascore.utils.mapping import FrozenDict
 from dascore.utils.misc import (
-    _dict_list_diffs,
-    _merge_tuples,
-    all_diffs_close_enough,
-    get_middle_value,
-    iterate,
-    separate_coord_info,
     to_str,
-    warn_or_raise,
 )
 from dascore.utils.models import (
     CommaSeparatedStr,
@@ -50,7 +34,7 @@ _coord_summary_suffixes = set(CoordSummary.model_fields)
 _coord_required = {"min", "max"}
 
 
-def _get_coords_dict(data_dict, fields):
+def _get_coords_dict(data_dict):
     """
     Add coords dict to data dict, pop out any coordinate attributes.
 
@@ -73,9 +57,6 @@ def _get_coords_dict(data_dict, fields):
     coord_info, new_attrs = separate_coord_info(
         data_dict, dims, required=("min", "max")
     )
-    if "dims" not in new_attrs and dims is not None:
-        new_attrs["dims"] = dims
-
     new_attrs["coords"] = {i: dc.core.CoordSummary(**v) for i, v in coord_info.items()}
     return new_attrs
 
@@ -157,7 +138,7 @@ class PatchAttrs(DascoreBaseModel):
     def parse_coord_attributes(cls, data: Any) -> Any:
         """Parse the coordinate attributes into coord dict."""
         if isinstance(data, dict):
-            data = _get_coords_dict(data, cls.model_fields)
+            data = _get_coords_dict(data)
             # add dims as coords if dims is not included.
             if "dims" not in data:
                 data["dims"] = tuple(data["coords"])
@@ -302,425 +283,3 @@ class PatchAttrs(DascoreBaseModel):
                 elif isinstance(start, float | np.floating):
                     out[step_name] = np.NaN
         return out
-
-
-@compose_docstring(conflict_desc=attr_conflict_description)
-def combine_patch_attrs(
-    model_list: Sequence[PatchAttrs],
-    coord_name: str | None = None,
-    conflicts: Literal["drop", "raise", "keep_first"] = "raise",
-    drop_attrs: Sequence[str] | None = None,
-    coord: None | BaseCoord = None,
-) -> PatchAttrs:
-    """
-    Merge Patch Attributes along a dimension.
-
-    Parameters
-    ----------
-    model_list
-        A list of models.
-    coord_name
-        The coordinate, usually a dimension coord, along which to merge.
-    conflicts
-        {conflict_desc}
-    drop_attrs
-        If provided, attributes which should be dropped.
-    coord
-        The coordinate for the new values of dim. This is provided as a
-        shortcut if it has already been computed.
-    """
-    # TODO this is a monstrosity! need to refactor.
-    eq_coord_fields = set(CoordSummary.model_fields) - {"min", "max", "step"}
-
-    def _get_merge_coords(model_dicts):
-        """Get the coordinates to merge together."""
-        # pop out coord to merge on
-        merge_coords = []
-        for mod in model_dicts:
-            coords = mod.get("coords", {})
-            maybe_coord = coords.pop(coord_name, None)
-            if maybe_coord is not None:
-                merge_coords.append(maybe_coord)
-        # dealing with empty coords
-        if not merge_coords:
-            return {}
-        return merge_coords
-
-    def _get_merge_dict_list(merge_coords):
-        """Get a list of {attrs: []}."""
-        out = defaultdict(list)
-        for mdict in merge_coords:
-            for key, val in mdict.items():
-                out[key].append(val)
-        return out
-
-    def _get_new_coord(model_dicts):
-        """Get the merged coord to set on all models."""
-        merge_coords = _get_merge_coords(model_dicts)
-        merge_dict_list = _get_merge_dict_list(merge_coords)
-        out = {}
-        # empy dicts
-        if not merge_dict_list:
-            return dict(merge_dict_list)
-        # all these should be equal else raise.
-        for key in eq_coord_fields:
-            if not len(vals := merge_dict_list[key]):
-                continue
-            if not all(vals[0] == x for x in vals):
-                msg = f"Cant merge patch attrs, key {key} not equal."
-                raise AttributeMergeError(msg)
-            out[key] = vals[0]
-        out["min"] = min(merge_dict_list["min"])
-        out["max"] = max(merge_dict_list["max"])
-        step = None
-        if all_diffs_close_enough(steps := merge_dict_list["step"]):
-            step = get_middle_value(steps)
-        out["step"] = step
-        return out
-
-    def _get_model_dict_list(mod_list):
-        """Get list of model_dict, merge along dim if specified."""
-        model_dicts = [
-            x.model_dump(exclude_defaults=True) if not isinstance(x, dict) else x
-            for x in mod_list
-        ]
-        # drop attributes specified.
-        if drop := set(iterate(drop_attrs)):
-            model_dicts = [
-                {i: v for i, v in x.items() if i not in drop} for x in model_dicts
-            ]
-        # no coordinate to merge on, just return dicts.
-        if not coord_name:
-            return model_dicts
-        # drop models which don't have required coords
-        model_dicts = [x for x in model_dicts if coord_name in x.get("coords", {})]
-        # coordinate can be determined from existing coords.
-        if coord is None:
-            new_coord = _get_new_coord(model_dicts)
-        else:  # or if one was passed just use its summary.
-            new_coord = coord.to_summary()
-        if new_coord:
-            for mod in model_dicts:
-                mod["coords"][coord_name] = new_coord
-        return model_dicts
-
-    def _replace_null_with_none(mod_dict_list):
-        """Because NaN != NaN we need to replace those values so == works."""
-        out = []
-        for mod in mod_dict_list:
-            out.append(
-                {
-                    i: (v if (isinstance(v, Sequence) or not pd.isnull(v)) else None)
-                    for i, v in mod.items()
-                }
-            )
-        return out
-
-    def _keep_eq(d1, d2):
-        """Keep only the values that are equal between d1/d2."""
-        out = {}
-        for i in set(d1) & set(d2):
-            if not d1[i] == d2[i]:
-                continue
-            out[i] = d1[i]
-        return out
-
-    def _handle_other_attrs(mod_dict_list):
-        """Check the other attributes and handle based on conflicts param."""
-        if conflicts == "keep_first":
-            return [dict(ChainMap(*mod_dict_list))]
-        no_null_ = _replace_null_with_none(mod_dict_list)
-        all_eq = all(no_null_[0] == x for x in no_null_[1:])
-        if all_eq:
-            return mod_dict_list
-        if conflicts == "raise":
-            # determine which keys are not equal to help debug.
-            uneq_keys = _dict_list_diffs(mod_dict_list)
-            msg = (
-                "Cannot merge models, the following non-dim attrs are not "
-                f"equal: {uneq_keys}. Consider setting the `conflict` or "
-                f"`attr_conflict` arguments for more flexibility in merging "
-                f"unequal coordinates."
-            )
-            raise AttributeMergeError(msg)
-        final_dict = reduce(_keep_eq, mod_dict_list)
-        return [final_dict]
-
-    mod_dict_list = _get_model_dict_list(model_list)
-    mod_dict_list = _handle_other_attrs(mod_dict_list)
-    first_class = model_list[0].__class__
-    cls = first_class if not first_class == dict else PatchAttrs
-    if not len(mod_dict_list):
-        msg = "Failed to combine patch attrs"
-        raise AttributeMergeError(msg)
-    return cls(**mod_dict_list[0])
-
-
-def check_dims(
-    patch1,
-    patch2,
-    check_behavior: WARN_LEVELS = "raise",
-    intersection: bool = False,
-) -> bool:
-    """
-    Return True if dimensions of two patches are equal.
-
-    Parameters
-    ----------
-    patch1
-        first patch
-    patch2
-        second patch
-    check_behavior
-        String with 'raise' will raise an error if incompatible,
-        'warn' will provide a warning, None will do nothing.
-    intersection
-        If True, allow any intersection of dimensions to pass. This is useful
-        when only broad-castablity needs to be checked. If false require dims
-        to be equal.
-    """
-    dims1, dims2 = patch1.dims, patch2.dims
-    dims_ok = True
-    if not intersection and patch1.dims == patch2.dims:
-        return True
-    dset1, dset2 = set(dims1), set(dims2)
-    if intersection and (dset1 | dset2):
-        return True
-    msg = (
-        "Patch dimensions are not compatible for merging."
-        f" Patch1 dims: {dims1}, Patch2 dims: {dims2}"
-    )
-    warn_or_raise(msg, exception=IncompatiblePatchError, behavior=check_behavior)
-    return dims_ok
-
-
-def check_coords(
-    patch1, patch2, check_behavior: WARN_LEVELS = "raise", dim_to_ignore=None
-) -> bool:
-    """
-    Return True if the coordinates of two patches are compatible, else False.
-
-    Parameters
-    ----------
-    patch1
-        patch 1
-    patch2
-        patch 2
-    check_behavior
-        String with 'raise' will raise an error if incompatible,
-        'warn' will provide a warning.
-    dim_to_ignore
-        None by default (all coordinates must be identical).
-        String specifying a dimension that differences in values,
-        but not shape, are allowed.
-    """
-    cm1 = patch1.coords
-    cm2 = patch2.coords
-    cset1, cset2 = set(cm1.coord_map), set(cm2.coord_map)
-    shared = cset1 & cset2
-    not_equal_coords = []
-    for coord in shared:
-        coord1 = cm1.coord_map[coord]
-        coord2 = cm2.coord_map[coord]
-        if coord1 == coord2:
-            # Straightforward case, coords are identical.
-            continue
-        elif coord == dim_to_ignore:
-            # If dimension that's ok to ignore value differences,
-            # check whether shape is the same.
-            if coord1.shape == coord2.shape:
-                continue
-            else:
-                not_equal_coords.append(coord)
-        else:
-            not_equal_coords.append(coord)
-    if not_equal_coords and len(shared):
-        msg = (
-            f"Patches are not compatible. The following shared coordinates "
-            f"are not equal: {coord}"
-        )
-        warn_or_raise(msg, exception=IncompatiblePatchError, behavior=check_behavior)
-        return False
-    return True
-
-
-def _merge_aligned_coords(cm1, cm2):
-    """Merge aligned coordinates removing non coords."""
-    assert cm1.dims == cm2.dims, "coordinates are not aligned"
-    out = {}
-    for name in set(cm1.coord_map) & set(cm2.coord_map):
-        coord1 = cm1.coord_map[name]
-        coord2 = cm2.coord_map[name]
-        # Coords already equal, just use first.
-        if coord1.approx_equal(coord2):
-            out[name] = coord1
-        # Deal with Non coords
-        non_count = sum([coord1._non_coord, coord2._non_coord])
-        if non_count == 1:
-            out[name] = coord1 if coord2._non_coord else coord2
-        elif non_count == 2:
-            out[name] = coord1 if coord1.size > coord2.size else coord2
-        assert name in out
-    return cm1.update(**out)
-
-
-def _merge_models(attrs1, attrs2, coord=None, attrs_to_ignore=DEFAULT_ATTRS_TO_IGNORE):
-    """Ensure models are equal in the right ways, merge together."""
-    no_comp_keys = set(attrs_to_ignore)
-    if attrs1 == attrs2:
-        return attrs1
-    dict1, dict2 = dict(attrs1), dict(attrs2)
-    if coord is not None:
-        new_coords = coord.to_summary_dict()
-        dict1["coords"], dict2["coords"] = new_coords, new_coords
-    else:
-        dict1.pop("coords"), dict2.pop("coords")
-    common_keys = set(dict1) & set(dict2)
-    ne_attrs = []
-    for key in common_keys:
-        if key in no_comp_keys:
-            continue
-        if dict2[key] != dict1[key]:
-            ne_attrs.append(key)
-    if ne_attrs:
-        msg = (
-            "Patches are not compatible because the following attributes "
-            f"are not equal. {ne_attrs}"
-        )
-        raise IncompatiblePatchError(msg)
-    return combine_patch_attrs([dict1, dict2], conflicts="keep_first")
-
-
-def merge_compatible_coords_attrs(
-    patch1: PatchType,
-    patch2: PatchType,
-    attrs_to_ignore=DEFAULT_ATTRS_TO_IGNORE,
-    dim_intersection: bool = False,
-    validate_coords: bool = True,
-) -> tuple[CoordManager, PatchAttrs]:
-    """
-    Merge the coordinates and attributes of patches or raise if incompatible.
-
-    The rules for compatibility are:
-
-    - All attrs must be equal other than those specified in attrs_to_ignore.
-    - Patches must share the same dimensions unless dim_intersection == True.
-    - All shared dimensional coordinates must be strictly equal
-    - If patches share a non-dimensional coordinate they must be equal.
-
-    Any coordinates or attributes contained by a single patch will be included
-    in the output.
-
-    Parameters
-    ----------
-    patch1
-        The first patch
-    patch2
-        The second patch
-    attrs_to_ignore
-        A sequence of attributes to not consider in equality. Only these
-        attributes from the first patch are kept in outputs.
-    dim_intersection
-        If True, merge if any dimensions overlap, else raise if all do not
-        overlap.
-    validate_coords
-        If True, ensure the coords are equal, else the responsibility for this
-        was handled upstream.
-    """
-
-    def _merge_coords(coords1, coords2):
-        out = {}
-        cmap1, cmap2 = coords1.coord_map, coords2.coord_map
-        coord_names = set(cmap1) | set(cmap2)
-        # fast path to update identical coordinates
-        if coord_names == set(cmap1):
-            return coords1
-        if coord_names == set(cmap2):
-            return coords2
-        # otherwise just squish coords from both managers together.
-        for name in coord_names:
-            coord = coords1 if name in coords1.coord_map else coords2
-            dims = coord.dim_map[name]
-            out[name] = (dims, coord.coord_map[name])
-        # Need to get coordinate that are in output, but preserve order.
-        dims = _merge_tuples(coords1.dims, coords2.dims)
-        return dc.core.coordmanager.get_coord_manager(out, dims=dims)
-
-    check_dims(patch1, patch2, intersection=dim_intersection)
-    if validate_coords:
-        check_coords(patch1, patch2)
-    coord1, coord2 = patch1.coords, patch2.coords
-    attrs1, attrs2 = patch1.attrs, patch2.attrs
-    coord_out = _merge_coords(coord1, coord2)
-    attrs = _merge_models(attrs1, attrs2, coord_out, attrs_to_ignore=attrs_to_ignore)
-    return coord_out, attrs
-
-
-def decompose_attrs(attr_list: Sequence[PatchAttrs], exclude=("history",)):
-    """Function to decompose attributes into series."""
-
-    def _get_uri_and_hash(model):
-        """Pop out this models uri and its hash."""
-        uri_key = "uri" if "uri" in model else "path"
-        assert uri_key in model, "all models must have uri or path"
-        uri = model.pop(uri_key)
-        uri_hash = hash(uri)
-        return uri, uri_hash
-
-    def _add_coords(coord_dict, dim_dict, coords, uri_hash, dims):
-        """Add coordinates to structure."""
-        for name, coord in coords.items():
-            out_dct = coord_dict if name not in dims else dim_dict
-            coord["index"] = uri_hash
-            coord["units"] = "" if coord["units"] is None else coord["units"]
-            # strip off the [] in datetimes/timedeltas. We will ensure the
-            # precision is always ns downstream.
-            if dtype := coord.get("dtype"):
-                coord["dtype"] = dtype.split("[")[0]
-            out_dct[name][coord["dtype"]].append(coord)
-
-    def _add_attrs(attr_dict, model, uri_hash):
-        """Add attrs to coordinate structure."""
-        for name, value in model.items():
-            if name.startswith("_"):
-                continue
-            entry = {"value": value, "index": uri_hash}
-            attr_dict[name][str(type(value))].append(entry)
-
-    def _to_df(dict_list):
-        """Convert a dict list to a dataframe."""
-        df = pd.DataFrame(dict_list).set_index("index").dropna(how="all")
-        # TODO may need to do something faster here, not sure.
-        empty_str = df == ""
-        null = pd.isnull(df)
-        to_keep = ~(empty_str | null).all(axis="columns")
-        return df[to_keep]
-
-    def _pandify(decom):
-        """Convert all the nested dicts to series."""
-        out = {}
-        for key, value in decom.items():
-            if isinstance(value, list):
-                df = _to_df(value)
-                if not df.empty:
-                    out[key] = df
-            else:
-                out[key] = _pandify(value)
-        return out
-
-    dumped = [x.model_dump(exclude=exclude) for x in attr_list]
-    out = {
-        "uri": [],
-        "attrs": defaultdict(lambda: defaultdict(list)),
-        "coords": defaultdict(lambda: defaultdict(list)),
-        "dims": defaultdict(lambda: defaultdict(list)),
-    }
-    for ind, model in enumerate(dumped):
-        uri, uri_hash = _get_uri_and_hash(model)
-        out["uri"].append({"uri": uri, "index": uri_hash})
-        dims = model.get("dims", "").split(",")
-        # need to handle coordinates and regular attrs differently
-        coords = model.pop("coords", {})
-        _add_coords(out["coords"], out["dims"], coords, uri_hash, dims)
-        _add_attrs(out["attrs"], model, uri_hash)
-    return _pandify(out)
