+++ pyproject.toml
@@ -160,7 +160,7 @@ target-version = "py310"
 fixable = ["ALL"]
 
 # List of codes to ignore
-ignore = ["D105", "D107", "D401", "D205", "D200"]
+ignore = ["D105", "D107", "D401", "D205", "D200", "D400"]
 
 [tool.ruff.mccabe]
 # Unlike Flake8, default to a complexity level of 10.
