+++ tests/test_utils/test_patch_utils.py
@@ -6,12 +6,14 @@ import pytest
 
 import dascore as dc
 from dascore.exceptions import (
+    IncompatiblePatchError,
     PatchAttributeError,
     PatchDimError,
 )
 from dascore.utils.patch import (
     align_patch_coords,
     get_dim_value_from_kwargs,
+    merge_compatible_coords_attrs,
 )
 
 
@@ -191,3 +193,60 @@ class TestAlignPatches:
         non_patch = random_patch.first("time")
         # Coords should be expanded in out2.
         out1, out2 = align_patch_coords(non_patch, random_patch)
+
+
+class TestMergeCompatibleCoordsAttrs:
+    """Tests for merging compatible attrs, coords."""
+
+    def test_simple(self, random_patch):
+        """Simple merge test."""
+        coords, attrs = merge_compatible_coords_attrs(random_patch, random_patch)
+        assert coords == random_patch.coords
+        assert attrs == random_patch.attrs
+
+    def test_incompatible_dims(self, random_patch):
+        """Ensure incompatible dims raises."""
+        new = random_patch.rename_coords(time="money")
+        match = "not compatible for merging"
+        with pytest.raises(IncompatiblePatchError, match=match):
+            merge_compatible_coords_attrs(random_patch, new)
+
+    def test_incompatible_coords(self, random_patch):
+        """Ensure an incompatible error is raised for coords that dont match."""
+        new_time = random_patch.attrs.time_max
+        new = random_patch.update_attrs(time_min=new_time)
+        match = "coordinates are not equal"
+        with pytest.raises(IncompatiblePatchError, match=match):
+            merge_compatible_coords_attrs(new, random_patch)
+
+    def test_incompatible_attrs(self, random_patch):
+        """Ensure if attrs are off an Error is raised."""
+        new = random_patch.update_attrs(network="TA")
+        match = "attributes are not equal"
+        with pytest.raises(IncompatiblePatchError, match=match):
+            merge_compatible_coords_attrs(new, random_patch)
+
+    def test_extra_coord(self, random_patch, random_patch_with_lat_lon):
+        """Extra coords on both patch should end up in the merged patch."""
+        new_coord = np.ones(random_patch.coord_shapes["time"])
+        pa1 = random_patch.update_coords(new_time=("time", new_coord))
+        pa2 = random_patch_with_lat_lon.update_coords(new_time=("time", new_coord))
+        expected = set(pa1.coords.coord_map) | set(pa2.coords.coord_map)
+        coords, attrs = merge_compatible_coords_attrs(pa1, pa2)
+        assert set(coords.coord_map) == expected
+        assert set(attrs.coords) == expected
+
+    def test_extra_attrs(self, random_patch):
+        """Ensure extra attributes are added to patch."""
+        patch = random_patch.update_attrs(new_attr=10)
+        coords, attrs = merge_compatible_coords_attrs(patch, random_patch)
+        assert attrs.get("new_attr") == 10
+
+    def test_different_dims(self, random_patch):
+        """Ensure we can merge patches which share some dims but not all."""
+        patch1 = random_patch
+        patch2 = random_patch.rename_coords(time="money")
+        coord, attrs = merge_compatible_coords_attrs(
+            patch1, patch2, dim_intersection=True
+        )
+        assert set(coord.dims) == (set(patch1.dims) | set(patch2.dims))
